requirements.txt
oci
ujson
opencv-python
Pillow